prayerbook-115 Відрефакторити код із використанням патерну MvvM або MVP, використати бібліотеку D... - in progress #115 - TextViewPresenter - in progress

+++ PrayerBook/app/src/main/java/com/tshevchuk/prayer/data/cache/InMemoryCacheManager.java
@@ -6,7 +6,7 @@ package com.tshevchuk.prayer.data.cache;
 public interface InMemoryCacheManager {
     void putCharSequence(String key, CharSequence text);
 
-    void putCharSequence(String key, CharSequence value, long size);
+    void putCharSequence(String key, CharSequence value, int size);
 
     CharSequence getCharSequence(String key);
 }
