+++ PrayerBook/app/src/main/java/com/tshevchuk/prayer/data/html_parser/HtmlParser.java
@@ -6,5 +6,5 @@ import android.support.v4.util.Pair;
  * Created by taras on 04.04.16.
  */
 public interface HtmlParser {
-    Pair<CharSequence, Long> parseHtml(String htmlToParse);
+    Pair<CharSequence, Integer> parseHtml(String htmlToParse);
 }
